Reorganized broadcast functions and added initial PLAY_TILE files
