+++ src/functions/onCreateGame.ts
@@ -5,13 +5,14 @@ import { Logger } from '../utils/Logger';
 import { WebSocketClient } from '../websocket/WebSocketClient';
 import { createGameResponse, successWebSocketResponse, failedWebSocketResponse } from '../websocket/createWSResponse';
 import { Game } from '../models/Game';
-import { broadcastGameUpdate, getConnectionIdsFromUsers } from '../websocket/broadcast';
 import { removeDynamoDocumentVersion } from '../dynamodb/dbHelper';
 import { LambdaEventBody, WebSocketAPIGatewayEvent } from '../types/event';
 import { LambdaEventBodyPayloadOptions } from '../types/payload';
 import { LambdaResponse } from '../types/response';
 import { GameStates } from '../types/states';
 import { getAllConnections, setGameIdForUser } from '../dynamodb/userDBService';
+import { broadcastGameUpdate } from '../websocket/broadcast/gameBroadcast';
+import { getConnectionIdsFromUsers } from '../utils/broadcastHelper';
 
 /**
  * Handler for creating a game.
