+++ src/functions/onDisconnect.ts
@@ -7,10 +7,11 @@ import { LambdaResponse } from '../types/response';
 import { removeUserFromGame } from '../dynamodb/gameDBService';
 import { User } from '../models/User';
 import { WebSocketClient } from '../websocket/WebSocketClient';
-import { broadcastUserUpdate, getConnectionIdsFromUsers } from '../websocket/broadcast';
 import { UserStates } from '../types/states';
 import { Game } from '../models/Game';
 import { sendUpdates } from './functionsHelper';
+import { broadcastUserUpdate } from '../websocket/broadcast/userBroadcast';
+import { getConnectionIdsFromUsers } from '../utils/broadcastHelper';
 
 /* ----------------------------------------------------------------------------
  * Handler Helper Functions
