+++ src/functions/onDrawTile.ts
@@ -5,7 +5,7 @@ import { Logger } from '../utils/Logger';
 import { LambdaEventBodyPayloadOptions } from '../types/payload';
 import { WebSocketClient } from '../websocket/WebSocketClient';
 import { response } from '../utils/responseHelper';
-import { broadcastDrawTileToUser } from '../websocket/broadcast';
+import { broadcastDrawTileToUser } from '../websocket/broadcast/gameBroadcast';
 
 /**
  * Handler for drawing a tile from the mahjong wall.
